refactor: [#1519] extract mod coordinator
