+++ packages/swarm-coordination-registry/src/swarm/mod.rs
@@ -1,0 +1,1 @@
+pub mod coordinator;
